refactor(userspace-dp): #6386 extract resolver_enqueue sibling
Move the #1769/#1912 per-key rate-limited neighbor-resolver enqueue
helper out of poll_descriptor/mod.rs into a new pub(super) sibling
resolver_enqueue.rs, plus its test module try_enqueue_resolver_tests
(loaded from the sibling via #[path]). The moved bare fn becomes
pub(super) and keeps its existing #[inline].

Pure code motion, attr-verbatim: no #[inline] added. Body byte-identical
to its prior location (verified with git diff --color-moved). The helper
is self-contained (afxdp-level deps via `use super::*;`), so no
cross-sibling import and no mod.rs prune are needed. mod.rs re-imports it
via `use resolver_enqueue::try_enqueue_resolver;` so every loop call site
is textually unchanged.

try_enqueue_resolver sat between the stranded poll_binding_process_descriptor
contract comment (#678/#946/#1826 area raw-pointer contract) and the fn
itself; removing it reunites that comment directly above
poll_binding_process_descriptor, completing the #6386 goal that mod.rs is
now "imports + the contract comment + the plan-killed (#4404) dispatch
core, nothing else".

Validation: cargo build (release) clean; cargo test --no-run clean;
moved body byte-identical modulo pub(super).
